fix: webpack plugin called twice
